Refactor Queue to EventQueue across the advanced-events module
- Updated class names and references from Queue to EventQueue in UML diagrams, domain entities, use cases, repositories, and controllers.
- Commented out the WebTopicController registration in the app configuration.
- Adjusted error messages and success responses in the QueueController to reflect the new EventQueue terminology.
- Modified the Agentry documentation to replace instances of Device with MobileDevice, including UML diagrams and API references.
- Added new library build for advanced-event-mesh-platform-service.

+++ agentry/source/uim/platform/agentry/presentation/rest/services/app_definition.d
@@ -29,7 +29,7 @@ struct AppDefinitionResponse {
     string tenantId;
     string id;
     string message;
-    int code;
+    ulong code;
     // string description;
 }
